+++ package.json
@@ -40,6 +40,7 @@
     "typescript-eslint": "^8.39.0",
     "vite": "^7.1.1",
     "vitest": "^3.2.4",
-    "vitest-browser-react": "^1.0.1"
+    "vitest-browser-react": "^1.0.1",
+    "vite-tsconfig-paths": "^5.1.4"
   }
 }
