+++ src/shared/components/Card.tsx
@@ -1,0 +1,40 @@
+import { forwardRef, HTMLAttributes } from "react"
+
+interface CardProps extends HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const Card = forwardRef<HTMLDivElement, CardProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`rounded-lg border bg-card text-card-foreground shadow-sm ${className}`} {...props} />
+))
+Card.displayName = "Card"
+
+// 카드 헤더 컴포넌트
+interface CardHeaderProps extends HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const CardHeader = forwardRef<HTMLDivElement, CardHeaderProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`flex flex-col space-y-1.5 p-6 ${className}`} {...props} />
+))
+CardHeader.displayName = "CardHeader"
+
+// 카드 타이틀 컴포넌트
+interface CardTitleProps extends HTMLAttributes<HTMLHeadingElement> {
+  className?: string
+}
+
+export const CardTitle = forwardRef<HTMLHeadingElement, CardTitleProps>(({ className, ...props }, ref) => (
+  <h3 ref={ref} className={`text-2xl font-semibold leading-none tracking-tight ${className}`} {...props} />
+))
+CardTitle.displayName = "CardTitle"
+
+// 카드 컨텐츠 컴포넌트
+interface CardContentProps extends HTMLAttributes<HTMLDivElement> {
+  className?: string
+}
+
+export const CardContent = forwardRef<HTMLDivElement, CardContentProps>(({ className, ...props }, ref) => (
+  <div ref={ref} className={`p-6 pt-0 ${className}`} {...props} />
+))
+CardContent.displayName = "CardContent"
