+++ tsconfig.app.json
@@ -19,7 +19,16 @@
     "noUnusedLocals": true,
     "noUnusedParameters": true,
     "noFallthroughCasesInSwitch": true,
-    "noUncheckedSideEffectImports": true
+    "noUncheckedSideEffectImports": true,
+
+    /* Paths */
+    "baseUrl": ".",
+    "paths": {
+      "/*": ["src/*"],
+      "shared/*": ["src/shared/*"],
+      "widgets/*": ["src/widgets/*"],
+      "pages/*": ["src/pages/*"]
+    }
   },
   "include": ["src"]
 }
